Updated openstack/openstack Project: openstack-infra/tripleo-ci  c1c549ae5ee532068c5cdbc28c8b23ebdc3f4509
Fix tuskar-ui disablement

enable_tuskar was only getting toggled to false in the sample
undercloud.conf which has no effect when the undercloud installer is
actually run. This change copies the sample file to the stack user's
home directory, and sets enable_tuskar=false there instead so it takes
effect.

Change-Id: I8a525a07d183c6a0b40d2ab11dd1148f1de93280