Updated openstack/openstack Project: openstack/nova  915e0d18b9cb9d8ff6d295fad19482d178d7655a
Add os-brick's scsi_id command to rootwrap

Fibre channel drivers are failing on calls from the
os-brick library to scsi_id.

This is a temporary fix until os-brick's rootwrap
filters can be pulled in. It is not intended or
desired that nova track all os-brick rootwrap filter
values. But until that mechanism is finalized, we
need necessary commands added here.

Change-Id: Ie5f1fbe0a2da171d80faefe3245df91aef8ce1ab
Partial-bug: #1479842