Updated openstack/openstack Project: openstack/python-openstackclient  535aac6c91b4df662647e5e35e681765da2c59f1
Extract log level from configuration file

Extract log_level from configuration file if the level was not
overridden by the command line option.  The default command line
option is 1 and there is no command line option to set the
verbose_level to 1, so if it is 1, it has not be set.

Change-Id: I1be04367c72f83c1181f92ca4c2c83165b66995c
Implements: blueprint logging-migration