Updated openstack/openstack Project: openstack/python-openstackclient  46696f5ed55c9313754e7019e727ee88fa0847ce
Create log configuration class

Configuration of logging gets triggered twice.  The first time it
uses the CLI options when the application is started and second
it uses the configuration file after that is read.  The state of
the logging needs to be saved from the first to the second time,
so I created a class.

Implements: blueprint logging-migration

Change-Id: I7b8d1a3b6fd128e98cafd7c16009c7b694a52146