Updated openstack/openstack Project: openstack/nova  40dbe07b145f8c6acc64784a7e3f258c5b7503e5
api: allow any scheduler hints

To help with v2.1 compat mode, we need to allow users to pass in any
scheduler hints, to allow for out of tree extensions.

This includes ensuring the soft additional properties validator does not
strip out extra properties when they are allowed by the schema.

As this Mail:
http://lists.openstack.org/pipermail/openstack-dev/2015-June/067996.html
pointed out the limit the scheduler-hints in the API is problematic. This
patch relax the validation of scheduler hints for v2.1 also.

Co-Authored-By: Ed Leafe <ed@leafe.com>
Co-Authored-By: Alex Xu <hejie.xu@intel.com>
Change-Id: I4c066075165f69d355a74979808fa0ad8d6546ab