Updated openstack/openstack Project: openstack/manila  3e0284b524ec6b068a5c8a59ec64e92b882e2ec0
Add SSL middleware to fix incorrect version host_url

The Manila API does not behave properly if it is behind an SSL
termination proxy. If this is the case, the host_url in version
resource are build using http protocol instead of https.

To handle to correct host_url with proxy, first, we should enable
X-Forwarded-Host in proxy so that Manila can get the right client
ip. Second, we should enable X-Forwarded-Proto, with the new SSL
middleware, Manila will get original protocol.

Change-Id: I2dac983481718e6d639453e3f03f41baf8cea56d
Closes-Bug: #1384379