Updated openstack/openstack Project: openstack/python-zaqarclient  8104ff488bf74dea82b4de4e88563c6b9e1299f3
Don't reuse request and transport instances

Re-using these instances was causing request failures when talking to
`uwsgi`. I still don't know the exact cause - probably something related
to the headers or even cached in the `http.Client` instance - but this
does fix the current issue and it also makes sense to have a new request
object for every request.

Change-Id: I30f78e048eea7105e4cced4530b165ad7a25d914