Updated openstack/openstack Project: openstack/ironic-python-agent  f683a783af2898c4753ad2fc2dce29a88a225976
Install "dmidecode" and "ipmitool" by default

The get_memory() method relies on dmidecode to get the total amount of
RAM present in the node but, but by default dmidecode is not installed
in the image.

Same for get_bmc_address() and ipmitool.

This patch fixes both cases by installing the "dmidecode" and "ipmitool"
packages when generating the CoreOS image of IPA.

Closes-Bug: 1494367
Change-Id: Ibded7bd533d604cf21a3b60afd396dc7b8b8750b