Updated openstack/openstack Project: openstack/nova  edfd7668938ccd3a38d8d6e4ec68fd7e85023f46
Fixes Hyper-V boot from volume fails when using ephemeral disk

Instance boot from volume fails in Hyper-V if the flavor has ephemeral
disk.

This happens because Hyper-V tries to attach both the ephemeral disk and
the boot volume to the same slot on the IDE controller.

This fixes the bug by making sure that the ephemeral disk will always
attach to the second slot.

Change-Id: Id6772b75683528db765241d21e25fba55358afb7
Closes-Bug: #1479129