Updated openstack/openstack Project: openstack/nova  640a61d1edbcebaef968776e97e8dacc86016ace
Handle nova-compute failure during a soft reboot

A soft reboot is soft in the sense that we let the instance
respond to ACPI events before shutdown. We still shutdown the
libvirt domain however.

Therefore, if nova-compute crashes having shutdown the instance
domain, but before starting it again, when nova-compute
restarts it will see an instance in a task_state indicating a
soft reboot, but with the power_state being shutdown. This was
unexpected and caused an instance.save() to crash out.

In those cases, change the task_state to one corresponding to
a hard reboot, and continue on.

Change-Id: Icdde0bc2e8c8c90ba20f48f010f230ae4d4dca54
Closes-Bug: #1072751