Updated openstack/openstack Project: openstack/nova  9b60e46a4cb4f669495677eb91302d9cd9585476
VMware: ensure that instance is deleted when volume is missing

In the event that the volume is missing, the  volume detachment fails,
we need to make sure that the instance deletion is done.

If this is not done then we will have the following:
1. Nova instance is deleted
2. Volume will still be in 'in-use' state
3. Backend VM will still be on the VC

In the event that a StorageError error is raised then we will
proceed to delete the instance.

All other exceptions will be raised so as to not delete the cinder
volume.

Change-Id: I99c59cb94d487322011071d24504dd9401879dba
Closes-bug: #1492121