Updated openstack/openstack Project: openstack/tempest  40a48b314ce944acef8e444f0a33d63aae4bdb30
Remove assertion checking rescoped issued_at

When using Fernet tokens with Keystone, all operations in this test potentially
happen within the same second, which mean the original issued_at time and the
rescoped issues_at time are the same, which fails an assertion. This isn't an
issue with uuid tokens because uuid tokens store full subsecond precision.
Fernet tokens also return subsecond format but the precision is always
truncated to .000000Z.  This is because the cryptography library determines the
issued_at time, not Keystone. Keystone uses the issued_at timestamp from Fernet
as the source of truth. This is why we should not have an assertion that checks
to make sure the original issued_at and the rescoped issued_at aren't the same.

Change-Id: I1b67486ce929cc5a2fc0fc36d14a8e6cf662fbfc
Closes-Bug: 1473567