Updated openstack/openstack Project: openstack/tempest  c8c6fcc89504d3f24f2ce3003ddebd4c7555737d
Added CONF.validation entry security_method

Added a config entry for defaulting the security_group and
security_group_rules instead of just hard-coding it if no
variable is provided.

This should still provide the same functionality as previously
written but give the option to be more flexible for
different environments that don't need the security_group
variables enabled.

Change-Id: Ic78805849d6dc1950f334d0142714b62a98ccf0a
Closes-Bug: #1486198