Updated openstack/openstack Project: openstack-dev/devstack  8ed89a8a560102cefa918b63e4f9ae6ba5386d52
Add trailing IDENTITY_API_VERSION to OS_AUTH_URL in swift_configure_tempurls

The python-keystoneclient requires a trailing /v<IDENTITY_API_VERSION> to
successfully authenticate, otherwise it fails with a 404 error due to a not
found resource.

This error showed up only when generating Swift tempurls, because the error was
raised when using python-swiftclient.

This change fixes this for python-swiftclient within devstack.

Change-Id: Ibe222d65162898db69acba076b5fe1cb3621fbc3
Closes-Bug: 1492216