Updated openstack/openstack Project: openstack/nova  91250f7bbfde9d4d7a4a6f116a6c97b217955a74
Add missing information to docstring

A recent change to the '_soft_validate_additional_properties()' method
in nova/api/validation/validators.py added new behaviors without
updating the docstring to reflect the new behaviors. Rather than block
that patch from merging, it was felt that a follow-up patch to fix the
docstring was preferable.

Closes-bug: 1494289
Change-Id: I3e32742a935b94eb211591bbf1e5695d5301101e