Updated openstack/openstack Project: openstack/neutron  19d769a833970939f26490bb4660390f0d872e79
Descheduling DVR routers when ports are unbound from VM

When a VM is deleted, the DVR port used by the VM could be unbound
from the compute node. When it is unbounded, it is no longer
in use on the node. Currently the unbind doesn't trigger a check
to determine whether the DVR router can be unscheduled from the
L3-agent running on the compute node. This patch makes the check
and unschedule the router, if necessary.

Closes-Bug: 1489184
Change-Id: I882e0682bfc7695b3b23e36eb4d7e35a5d19748e