Updated openstack/openstack Project: openstack/nova  8d0bdde3693cca44ca2177086322399dc53b1276
Sorting and pagination params used as filters

While retrieving servers, the sort and pagination query string
parameters are treated as search options.

These parameters are passed down to the DB layer and eventually
filtered out when an AttributeError is caught because they do not
exist on the Instance model.

This change set removes these parameters at the API layer.

Closes-Bug: #1446638

Change-Id: I32c92649a2d0346e13c2013bfaa423aaec3b1b67