Updated openstack/openstack Project: openstack/horizon  2279371e0b4c1e798139ef6d70a4f509140290dc
Launch instance create volume fails

From both launch instances, if selecting create volume the value passed
for delete_on_termination is converted to an int. This causes launch
instance to fail with an error like:

Error: Invalid input for field/attribute delete_on_termination. Value:
0. 0 is not of type 'boolean', 'string' (HTTP 400) (Request-ID:
req-aee15975-0351-4477-9007-f8fd090bcc49)

This is returned from the nova API. The only reference I can find for
the value type is:
https://github.com/openstack/nova/blob/master/nova/objects/block_device.py#L75

This change leaves the value as a boolean in both workflows and converts
the tests appropriately.

Closes-Bug: #1494067
Change-Id: I54591269eafe5b89b49a13c29c53611c100d600e