Updated openstack/openstack Project: openstack/cinder  3788754d5eb29af6ecb4f90067c93c402866ee44
SolidFire provider_id for snapshots on init

We've added the provider_id field to the database for both
volumes and snapshots.  This however couldn't be used because
there was no upgrade/migration path included for backends to
populate these fields (ie upgrade from K to L).

This patch fixes the provider_id columns for snapshots and volumes
both, and updates these fields on backend init.  This fixes the
issue of not being able to "rely" on the database entries, and it
also fixes some issues with inefficiencies in how we were deleting
resources.

Now we can just use the provider_id info from the reference passed
in on the delete calls.

Change-Id: I9de02e9d12aa23ddf4e1fbb1e0efab2c1805fdb1
Closes-Bug: #bug/1491489