Updated openstack/openstack Project: openstack/nova  31ef4f4b8339e4e46ba680a4cc13e90649445c60
Add version 3.0 of conductor RPC interface

This commit adds a new major revision to the rpc interface used to
talk to the conductor service. The old version will be retained in
Liberty for Kilo rolling upgrade capability. 2.x can be dropped as soon
as Mitaka development opens.

When 2.x gets dropped, we will be able to remove the following methods
from the conductor manager, as well as all other supporting code for
these methods throughout the conductor:

 - instance_update()
 - instance_get_by_uuid()
 - instance_get_all_by_host()
 - migration_get_in_progress_by_host_and_node()
 - aggregate_host_add()
 - aggregate_host_delete()
 - aggregate_metadata_get_by_host()
 - bw_usage_update()
 - agent_build_get_by_triple()
 - block_device_mapping_update_or_create()
 - block_device_mapping_get_all_by_instance()
 - instance_get_all_by_filters()
 - instance_get_active_by_window()
 - instance_destroy()
 - instance_fault_create()
 - vol_usage_update()
 - service_get_all_by()
 - action_event_start()
 - action_event_finish()
 - service_create()
 - service_destroy()
 - compute_node_create()
 - compute_node_update()
 - compute_node_delete()
 - service_update()
 - task_log_get()
 - task_log_begin_task()
 - task_log_end_task()
 - notify_usage_exists()
 - security_groups_trigger_handler()
 - security_groups_trigger_members_refresh()
 - network_migrate_instance_start()
 - network_migrate_instance_finish()
 - quota_commit()
 - quota_rollback()
 - get_ec2_ids()
 - compute_unrescue()
 - object_class_action()
 - object_backport()

Note that the client side will only send 3.0 messages, but that's
fine.  We expect the nova-conductor service to get updated at the same
time as, or before anything that uses this client code.

For more information on this type of change, see:
https://wiki.openstack.org/wiki/RpcMajorVersionUpdates

Related to blueprint liberty-bump-object-and-rpcapi-versions

Change-Id: Ia976c7f3362df9d91b2e7d23f0465b8d57149714