Updated openstack/openstack Project: openstack/heat  f026f4c322762a5d8c83fb4c0777a26024c4f3af
Set resource status to FAILED when engine is down

When resource is in status IN_PROGRESS and engine service
went down, user can't operate(such as update) the resource
anymore even if the engine service is back again.
This patch set the resources to FAILED status when
the engine starting.

Change-Id: Ib49895d45d727e117af5dbdcf84d58b196991274
Closes-Bug: #1487288