Updated openstack/openstack Project: openstack/horizon  0f39dea2c7e722781fe3374abbaada781488c2cc
Fix loading of angular in jquery modals

This patch fixes the invocation of the loadAngular method
which started failing (silently) when moved out of an
angular module context (variables it required were not
defined.

This patch also removes the try/catch that was silently
squashing errors, and adds documentation in the code
to explain what the loadAngular function's purpose
is.

Change-Id: I95382e89b5e8098af596319730328480c85e324d
Closes-Bug: 1494149