Updated openstack/openstack Project: openstack/tripleo-incubator  dbd7539c30e2fec27c70f64b8b9c555c213e6831
Do not exit when $TRIPLEO_ROOT is not defined

They are two different processes consuming this file:

* devtest
* instack

Devtest does need the $TRIPLEO_ROOT variable to be set up, but instack
does not. In order to allow this file to be usable by both processes, a
WARNING is dipslayed instead of ERRORing.

Change-Id: I66646f12793e2923c9bb25dca130d90477b94779