Updated openstack/openstack Project: openstack/nova  9a3a0f31777b68f85c8542fc522d5d449e6ff91e
Set vif and allocated when associating fixed ip

Follow on to changes I84fa2f0926f719bff33ac43b39e161a4eb412f44 and
I8cf5483982085da57ee470fa2753b0d0aebc12b3.

This change removes a race when associating fixed IPs and pools.
Originally, the fixed IPs were associated, and
"virtual_interface_id" was updated and saved to the DB at a later time.
This value is now sent into associate(_pool) and are tunneled down
to the DB on associate, so now the value is saved to the DB at
association.

Change-Id: Ic37d96deba3a5ef51594733cc16091173a9655a4
Closes-Bug: #1249065