Updated openstack/openstack Project: openstack/manila  c7e924848043881361132a41766c3749838e0d92
Enhance base driver checking if a method was implemented

The base ShareDriver "_has_redefined_driver_methods" method
checks if a stub method was implemented by comparing it to the
super class method.

This can cause issues if those methods are inherited, being
actually implemented in the super class, which does not have to be
the base ShareDriver class (for example, we could have common
classes, inherited by multiple drivers).

This patch fixes this by comparing the methods being tested with
the ones from the ShareDriver class.

Also, the create_share_from_snapshot method was removed from the
Huawei driver as it was actually just raising NotImplementedError.
This was misleading for our way of checking if a driver supports
snapshotting.

Closes-Bug: #1492311

Change-Id: I0369d086234053e5495a35db5034f7344c6c323a