Updated openstack/openstack Project: openstack/nova  efd7688d7fef2edbe0d4748e86095904cd8ce96b
Remove/deprecate unused parts of the compute node object

This change remove unused parts of the compute node object related to
its old connection with the service object. Unused remotable methods
that can't be removed until the next major version bump are left with
comments to do so when that happens.

Change-Id: Ia4bfd370a31c5b9286676b1f9440eeb420a60bec