Updated openstack/openstack Project: openstack/glance  1bc7b736def5ebb7b1fb964c3b329bbff2e465a7
Don't return 300 when requesting /versions

Whenever the versions resource is used, it'll reply with a 300 status
code. This is the desired behavior when requesting versions that we
don't support and/or non-existent.

However, when the user requests specifically for the `/versions`
endpoint, Glance shouldn't reply with a 300 since there's just 1 choice
for such endpoint.

APIImpact
Closes-bug: #1491350

Change-Id: I23e3780e5d1b609753ffef59418115f45ab45fe7