Updated openstack/openstack Project: openstack/manila  88e0465214d51069940827a709900f3e64ca53dc
Remove HDS SOP driver due to lack of CI

The Manila community set a deadline of 3 Sep 2015 for all drivers
to have reporting CI systems, and there is no CI system for the
HDS SOP driver. As agreed, the driver will be removed until such
time that a CI system exists.

Change-Id: I628fbcbbef42c7a215e0cca027c31f88ee6740aa