Updated openstack/openstack Project: openstack/cinder  0f6d84dceaa729c78e3e7222de774b2558a64754
backup init_host cleanup exception handling

Current cleanup of leftover backups, volumes, and temporary
volumes and snapshots is not resilient in the face of exceptions
such as failures in detach operations.

This commit adds appropriate try/except logic to this section
of code, thereby ensuring that the backup service can start even
if exceptions are encountered when attempting to cleanup detritus
from earlier failed backup operations.

Change-Id: Ie8d2107ebd4d86dfe5664c35687e8cbe58cfad9d
Closes-bug: 1485295