Updated openstack/openstack Project: openstack/horizon  da3dc00e69ac3e229a7d1af911f17abb370e1561
Fix ng Launch Instance button in Network Topology

This cleans up the anchor tag on the Network Topology page that is
used to open the angular Launch Instance modal. Basically it adds
the href to make it accessible, removes the ajax-modal class so the
modal opens properly, and removes the dismissUrl since dismissing
the modal should leave the user where they are.

Closes-Bug: #1494503
Change-Id: Ia6caf8fd4f0eceedf3046125780a3e013eacf0f1