Updated openstack/openstack Project: openstack-dev/devstack  d95ec812e032a45e3978c75b585963207ef66b5d
move back to editable install for oslo

Now that we don't have namespace packages any more, editable installs
should be fine. This also means that we apply constraints to these
libraries during installation, which is important for future testing.

This is needed in order to be able to easily sanity check
LIBS_FROM_GIT, as then all libs installed from git will have pip urls
with git in them.

Change-Id: I46c3b8f943b97f912eccc7278e3e033ae67e7e31