Updated openstack/openstack Project: openstack/manila  75fde5a26f1a6fd54c075edfe75de06ee131bb52
Remove support for 'latest' microversion

Originally copied from the Nova implementation of microversions,
we have come to realize that being able to specify 'latest' for the
microversion is rarely useful and usually dangerous.

Allowing clients to request the "latest" microversion guarantees
undefined (and likely broken) behavior in every situation where a
client talks to a server that is newer than it.

Change-Id: I3a5ba0d45804e567aba710a20628539b0c284499
Closes-Bug: 1489583