Updated openstack/openstack Project: openstack/neutron  01939adc4c429c993ddb6004b36a4d4b05f52cc7
ovs: don't use ARP responder for IPv6 addresses

ARP does not support IPv6 addresses, so when we try to apply the flow, it
fails, with all other flows deferred for the same transaction. It results in
random flow breakages, depending on the order of the bad flow in the
transaction.

Change-Id: I0ecf167653e5a7d0916e091e05050406a026a1e2
Co-Authored-By: Thomas Carroll <Thomas.Carroll@pnnl.gov>
Closes-Bug: #1477253