Updated openstack/openstack Project: openstack/neutron-vpnaas  582e682453d04c867ac03d82306f9bfdac51c50e
Support for VPN functional tests on Neutron commits

Currently, when the VPN functional tests run, it uses tox_install.sh to install
neutron code (as well as other packages, as passed in).

For functional tests invoked via the post_test_hook.sh script, tox_install.sh
would try to "import neutron", and if successful, then it is assumed that
Neutron is already installed. Otherwise a check is done to see if it can pip
install using the cache from Zuul cloner. For a local test run, tox_install.sh
would pip install Neutron from git.openstack.org.

If, however, one invokes the new Neutron experimental job that runs the VPN
functional tests on a Neutron patch set, tox_install.sh was determining that
Neutron was already installed. This was using the upstream version of Neutron
in /tmp/openstack, instead of the changeset code in /opt/stack/new.

This change attempts to fix this case, but first checking to see if the
Neutron directory exists at /opt/stack/new ($HOME/neutron actually). If it
does exist, that is pip installed for the VPN functional tests. Otherwise,
it does the same checks as explained above.

The end goal here is to be able to detect cross project breakages, by using
the (currently experimental) gate-neutron-vpnaas-dsvm-functional-sswan.

Change-Id: Ifa8ef994a571c72236a4a22a4eeb27496202db2b