Updated openstack/openstack Project: openstack/horizon  015fb6d94a1e8b1595bcc3ec81b8c9fe656c64b8
Fix a bug in "Add Interface" form modal

When Add Interface for a router that it's name is blank,
it cast error "This field is required" beside the router
name field. So there should display the router's name_or_id
in "Add Interface" form modal, and this is consistent with
the "Edit Router". This bug also exists in network topology
page. This patch fix the bug.

Change-Id: I0471f76b05eba9f546916cc6fe50134c88136098
Closes-Bug: #1493725