Updated openstack/openstack Project: openstack/neutron  6a845249d0bb004141b63c74ce80f6f68ae4a250
Reservations: Don't count usage if resource is unlimited

If a resource is unlimited (ie: limit<0) then there is no need
to verify headroom for it. This also means that there no need for
counting it; therefore it is possible to save some DB operations
by skipping the count phase.

Change-Id: Ibe9ca8a1c29fb8ba12df187c25f8f9515968a54d
Related-blueprint: better-quotas